Fix resolving remote type metadata documents
