+++ shared/src/commonMain/kotlin/at/asitplus/wallet/app/common/CredentialMetadataDisplayNames.kt
@@ -1,7 +1,6 @@
 package at.asitplus.wallet.app.common
 
-import kotlinx.atomicfu.atomic
-import kotlinx.atomicfu.update
+import androidx.compose.runtime.snapshots.SnapshotStateMap
 
 /**
  * In-memory lookup of human-readable credential type names taken from resolved SD-JWT VC Type Metadata, keyed by the
@@ -10,13 +9,16 @@ import kotlinx.atomicfu.update
  * Schemes resolved from remote type metadata are [at.asitplus.wallet.lib.data.ExtractedSdJwtCredentialScheme] etc.,
  * which do not carry the type-level display name — so without this, such credentials would show their bare `vct`.
  * Populated by [PersistentCachingCredentialMetadataRegistry] as metadata is resolved.
+ *
+ * Backed by a [SnapshotStateMap] so a [get] made during composition recomposes once the name is resolved (metadata is
+ * fetched asynchronously, so the value often arrives after the first render). Thread-safe for the background writes.
  */
 object CredentialMetadataDisplayNames {
-    private val names = atomic(mapOf<String, String>())
+    private val names = SnapshotStateMap<String, String>()
 
     operator fun set(identifier: String, displayName: String) {
-        names.update { if (it[identifier] == displayName) it else it + (identifier to displayName) }
+        names[identifier] = displayName
     }
 
-    operator fun get(identifier: String?): String? = identifier?.let { names.value[it] }
+    operator fun get(identifier: String?): String? = identifier?.let { names[it] }
 }
